Breaking API changes: Moving to objects from references.
I have modified mira a bit so it is more consistent. I removed usage
of pointers where it shouldn't have even been there. Codebase is more
consistent as now we just pass in structs. With using pointers we had
weird artifacts due to GC and just pointer gobbling. This should make
usage simpler. As this is a breaking update, the version bump will go
into primary. Tag 3.0.0 is incoming.

Signed-off-by: Sandy <thecsw@ku.edu>